requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
aiohttp
python-dotenv
cryptography
redis
playwright
google-cloud-run
google-cloud-monitoring
google-cloud-logging
google-cloud-translate
google-cloud-firestore
firebase-admin
google-auth
google-auth-oauthlib
google-adk
psutil
langdetect
requests
websockets
Pillow
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
pytest-mock
pytest-cov
httpx
slowapi
structlog
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-monitoring
browser-use